PythonProjects/Coding_Challenge_Generator/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api>=5.0.2",
    "fastapi>=0.125.0",
    "ollama>=0.3.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "svix>=1.86.0",
    "uvicorn>=0.41.0",
]
